+++ framework/src/main/java/org/moqui/context/TransactionFacade.java
@@ -121,4 +121,5 @@ public interface TransactionFacade {
 
     void initTransactionCache();
     boolean isTransactionCacheActive();
+    void flushAndDisableTransactionCache();
 }